Issue 2605 Move rule expanding logic to the separate class
